Better solution for dealing with calls from get.php
Initial solution does not seem to work for us because the path Autoload.php in in depends on the .modman directory. This checks to see if calling file is get.php and then uses the local variable $bp that get.php uses to hold the base path. Yes, I realize global variables are bad. 