Fix warnings when the rename dest dir is missing
In developer mode, warnings are turned into exceptions.
The rename call was not verifying that the destination directory existed and
causing warnings/exceptions in cases where the dest directory was missing.
This checks for the existance of the dest directory before trying to create and move the file.
Additionally, the rename call is now silenced to prevent additional warnings in the destructor.